feat: migrate region component
