--- src/region/index.js
@@ -1,3 +1,0 @@
-import Region from './Region';
-
-export default Region;
