+++ src/region/index.tsx
@@ -1,0 +1,5 @@
+import Region, { RegionProps } from './Region';
+
+export type { RegionProps };
+
+export default Region;
